integrate AR
