+++ pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   rive: ^0.9.1
   curved_labeled_navigation_bar: ^2.0.2
   animated_text_kit: ^4.2.2
+  arkit_plugin: ^1.0.7
 
 dev_dependencies:
   flutter_test:
